Refactor: 댓글 목록 조회 api endPoint 변경
